Lots of changes
Added particles to put over kill barriers, everyone will need to go thru and do that. Added particles to signify if a platform is active or not. Made bullets bigger so they're more readable for the player, also added particles to those. Fixed platforms not interacting correctly with sprite masks in levels 2, 4, and 5.